Skill ut regresjon. Legg til overgangstester
