+++ src/test/java/no/nav/foreldrepenger/stønadskonto/regelmodell/MinsterettTest.java
@@ -48,29 +48,6 @@ class MinsterettTest {
     }
 
     @Test
-    void wlb_bfhr_mor_ufør_flerbarn_før_utvidelse() {
-        var grunnlag80 = new BeregnKontoerGrunnlag.Builder()
-            .antallBarn(2)
-            .morHarUføretrygd(true)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(ETTER_WLB_1)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-        fellesassert(grunnlag80, 10, 0, 0, 105 + 95, 0);
-
-        var grunnlag100 = new BeregnKontoerGrunnlag.Builder()
-            .antallBarn(4)
-            .morHarUføretrygd(true)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(ETTER_WLB_1)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag100, 10, 0, 0, 230 + 75, 0);
-    }
-
-    @Test
     void wlb_bfhr_mor_ufør_flerbarn() {
         var grunnlag80 = new BeregnKontoerGrunnlag.Builder()
             .antallBarn(2)
@@ -92,18 +69,6 @@ class MinsterettTest {
             .build();
         fellesassert(grunnlag100, 10, 0, 0, 230 + 75, 0);
     }
-
-    @Test
-    void wlb_begge_rett_før_utvidelse() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .fødselsdato(ETTER_WLB_1)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .brukerRolle(Brukerrolle.FAR)
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .build();
-        fellesassert(grunnlag, 10, 0, 0, 0, 0);
-    }
-
     @Test
     void wlb_begge_rett() {
         var grunnlag = new BeregnKontoerGrunnlag.Builder()
@@ -113,17 +78,6 @@ class MinsterettTest {
             .rettighetType(Rettighetstype.BEGGE_RETT)
             .build();
         fellesassert(grunnlag, 10, 0, 0, 0, 0);
-    }
-
-    @Test
-    void wlb_bfhr_før_utvidelse() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .fødselsdato(ETTER_WLB_1)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .build();
-        fellesassert(grunnlag, 10, 0, 0, 40, 0);
     }
 
     @Test
@@ -143,7 +97,7 @@ class MinsterettTest {
             .rettighetType(Rettighetstype.ALENEOMSORG)
             .brukerRolle(Brukerrolle.FAR)
             .morHarUføretrygd(true)
-            .fødselsdato(ETTER_WLB_1)
+            .fødselsdato(ETTER_WLB_2)
             .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
             .build();
         fellesassert(grunnlag, 10, 0, 0, 0, 0);
@@ -269,28 +223,7 @@ class MinsterettTest {
     }
 
     @Test
-    void regeldato_bfhr() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .regelvalgsdato(FØR_WLB)
-            .fødselsdato(ETTER_WLB_1)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .build();
-        fellesassert(grunnlag, 0, 0, 0, 0, 0);
-
-        grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .regelvalgsdato(ETTER_WLB_1.minusMonths(1))
-            .fødselsdato(ETTER_WLB_1)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .build();
-        fellesassert(grunnlag, 10, 0, 0, 40, 0);
-    }
-
-    @Test
-    void regeldato_bfhr_2024_før() {
+    void regeldato_bfhr_2024_før_etter() {
         var grunnlag = new BeregnKontoerGrunnlag.Builder()
             .regelvalgsdato(ETTER_WLB_1)
             .fødselsdato(ETTER_WLB_2)
@@ -299,11 +232,8 @@ class MinsterettTest {
             .brukerRolle(Brukerrolle.FAR)
             .build();
         fellesassert(grunnlag, 10, 0, 0, 40, 0);
-    }
 
-    @Test
-    void regeldato_bfhr_2024_etter() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
+        grunnlag = new BeregnKontoerGrunnlag.Builder()
             .regelvalgsdato(ETTER_WLB_2)
             .fødselsdato(ETTER_WLB_2)
             .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
@@ -372,131 +302,6 @@ class MinsterettTest {
             .rettighetType(Rettighetstype.BEGGE_RETT)
             .brukerRolle(Brukerrolle.FAR)
             .omsorgsovertakelseDato(ETTER_WLB_1)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag, 0, 0, 0, 0, 0);
-    }
-
-    @Test
-    void gammel_bfhr_mor_ufør() {
-        var grunnlag80 = new BeregnKontoerGrunnlag.Builder()
-            .morHarUføretrygd(true)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(FØR_WLB)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-        fellesassert(grunnlag80, 0, 0, 0, 0, 95);
-
-        var grunnlag100 = new BeregnKontoerGrunnlag.Builder()
-            .morHarUføretrygd(true)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(FØR_WLB)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag100, 0, 0, 0, 0, 75);
-    }
-
-    @Test
-    void gammel_bfhr_mor_ufør_flerbarn() {
-        var grunnlag80 = new BeregnKontoerGrunnlag.Builder()
-            .antallBarn(2)
-            .morHarUføretrygd(true)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(FØR_WLB)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-        fellesassert(grunnlag80, 0, 0, 0, 0, 95);
-
-        var grunnlag100 = new BeregnKontoerGrunnlag.Builder()
-            .antallBarn(4)
-            .morHarUføretrygd(true)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(FØR_WLB)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag100, 0, 0, 0, 0, 75);
-    }
-
-    @Test
-    void gammel_begge_rett() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(FØR_WLB)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag, 0, 0, 0, 0, 0);
-    }
-
-    @Test
-    void gammel_bfhr() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(FØR_WLB)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag, 0, 0, 0, 0, 0);
-    }
-
-    @Test
-    void gammel_både_bfhr_og_alenesomsorg_skal_tolkes_som_aleneomsorg() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .rettighetType(Rettighetstype.ALENEOMSORG)
-            .brukerRolle(Brukerrolle.FAR)
-            .morHarUføretrygd(true)
-            .fødselsdato(FØR_WLB)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-        fellesassert(grunnlag, 0, 0, 0, 0, 0);
-    }
-
-    @Test
-    void gammel_totette_gir_ikke_dager() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .omsorgsovertakelseDato(FØR_WLB)
-            .familieHendelseDatoNesteSak(FØR_WLB.plusWeeks(40))
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag, 0, 0, 0, 0, 0);
-    }
-
-    @Test
-    void gammel_ikke_totette_gir_ikke_dager() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .fødselsdato(FØR_WLB)
-            //Barna ikke tett nok til minsteretten
-            .familieHendelseDatoNesteSak(FØR_WLB.plusWeeks(50))
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag, 0, 0, 0, 0, 0);
-    }
-
-    @Test
-    void gammel_bfhr_mor_rett_i_eøs_skal_ikke_få_generell_minsterett() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .fødselsdato(FØR_WLB)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-        fellesassert(grunnlag, 0, 0, 0, 0, 0);
-    }
-
-    @Test
-    void gammel_begge_rett_adopsjon_gir_ikke_dager_rundt_fødsel() {
-        var grunnlag = new BeregnKontoerGrunnlag.Builder()
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .omsorgsovertakelseDato(FØR_WLB)
             .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
             .build();
         fellesassert(grunnlag, 0, 0, 0, 0, 0);
