+++ src/test/java/no/nav/foreldrepenger/stønadskonto/regelmodell/StønadskontoRegelOrkestreringTest.java
@@ -32,8 +32,7 @@ import no.nav.foreldrepenger.stønadskonto.regelmodell.rettighet.PrematurukerUti
 
 class StønadskontoRegelOrkestreringTest {
 
-    private static final LocalDate FØR_WLB = LocalDate.of(2022, Month.JULY, 1);
-    private static final LocalDate ETTER_WLB_1 = LocalDate.of(2022, Month.DECEMBER,1);
+    private static final LocalDate PREMATUR = LocalDate.of(2019, Month.JULY, 1);
     private static final LocalDate ETTER_WLB_2 = LocalDate.of(2024, Month.DECEMBER,1);
 
     private final StønadskontoRegelOrkestrering stønadskontoRegelOrkestrering = new StønadskontoRegelOrkestrering();
@@ -110,10 +109,9 @@ class StønadskontoRegelOrkestreringTest {
 
     @Test
     void skal_legge_til_prematurdager_på_fellesperiode() {
-        var fødselsdato = LocalDate.of(2019, 7, 1);
         var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(fødselsdato)
-            .termindato(fødselsdato.plusWeeks(7).plusDays(4))
+            .fødselsdato(PREMATUR)
+            .termindato(PREMATUR.plusWeeks(7).plusDays(4))
             .antallBarn(1)
             .rettighetType(Rettighetstype.BEGGE_RETT)
             .brukerRolle(Brukerrolle.MOR)
@@ -134,10 +132,9 @@ class StønadskontoRegelOrkestreringTest {
 
     @Test
     void skal_legge_til_prematurdager_på_foreldrepenger() {
-        var fødselsdato = LocalDate.of(2019, 7, 1);
         var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(fødselsdato)
-            .termindato(fødselsdato.plusWeeks(8))
+            .fødselsdato(PREMATUR)
+            .termindato(PREMATUR.plusWeeks(8))
             .antallBarn(1)
             .rettighetType(Rettighetstype.ALENEOMSORG)
             .brukerRolle(Brukerrolle.FAR)
@@ -153,10 +150,9 @@ class StønadskontoRegelOrkestreringTest {
 
     @Test
     void skal_ikke_legge_til_prematurdager_på_flerbarnsdager() {
-        var fødselsdato = LocalDate.of(2019, 7, 1);
         var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(fødselsdato)
-            .termindato(fødselsdato.plusWeeks(7).plusDays(1))
+            .fødselsdato(PREMATUR)
+            .termindato(PREMATUR.plusWeeks(7).plusDays(1))
             .antallBarn(2)
             .rettighetType(Rettighetstype.BEGGE_RETT)
             .brukerRolle(Brukerrolle.MOR)
@@ -191,34 +187,6 @@ class StønadskontoRegelOrkestreringTest {
     }
 
     /*
-       Totale stønadskonto: - 59 uker (295 stønadsdager)
-       Stønadskonto fordeler seg slik:
-       - Fellesperiode: 18 uker (90 stønadsdager)
-       - Fedrekovte: 19 uker (95 stønadsdager)
-       - Mødrekvote: 19 uker (95 stønadsdager)
-       - Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
-   */
-    @Test
-    void fødsel_begge_har_rett_og_omsorg_dekningsgrad_80() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(1)
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(5)
-            .containsEntry(FELLESPERIODE, 90)
-            .containsEntry(FEDREKVOTE, 95)
-            .containsEntry(MØDREKVOTE, 95)
-            .containsEntry(FORELDREPENGER_FØR_FØDSEL, 15)
-            .containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
        Totale stønadskonto: - 61,2 uker (306 stønadsdager)
        Stønadskonto fordeler seg slik:
        - Fellesperiode: 20,2 uker (101 stønadsdager)
@@ -227,7 +195,7 @@ class StønadskontoRegelOrkestreringTest {
        - Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
     */
     @Test
-    void fødsel_begge_har_rett_og_omsorg_dekningsgrad_80_før_utvidelse() {
+    void fødsel_begge_har_rett_og_omsorg_dekningsgrad_80() {
         var grunnlag = BeregnKontoerGrunnlag.builder()
             .fødselsdato(ETTER_WLB_2)
             .antallBarn(1)
@@ -246,28 +214,6 @@ class StønadskontoRegelOrkestreringTest {
             .containsEntry(FAR_RUNDT_FØDSEL, 10);
     }
 
-
-    /*
-        Totale stønadskonto: - 56 uker (280 stønadsdager)
-        Stønadskonto fordeler seg slik:
-        - Fellesperiode: 18 uker (90 stønadsdager)
-        - Fedrekovte: 19 uker (95 stønadsdager)
-        - Mødrekvote: 19 uker (95 stønadsdager)
-    */
-    @Test
-    void adopsjon_begge_har_rett_og_omsorg_dekningsgrad_80_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .omsorgsovertakelseDato(ETTER_WLB_1)
-            .antallBarn(1)
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FELLESPERIODE, 90).containsEntry(FEDREKVOTE, 95).containsEntry(MØDREKVOTE, 95);
-    }
 
     /*
         Totale stønadskonto: - 56 uker (280 stønadsdager)
@@ -351,37 +297,7 @@ class StønadskontoRegelOrkestreringTest {
     /*
      Totale stønadskonto: - 59 uker (295 stønadsdager)
      Stønadskonto fordeler seg slik:
-        - Fellesperiode: 18 uker (90 stønadsdager) + 56 uker (280 stønadsdager)
-        - Fedrekovte: 19 uker (95 stønadsdager)
-        - Mødrekvote: 19 uker (95 stønadsdager)
-        - Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
-    */
-    @Test
-    void fødsel_begge_har_rett_og_omsorg_dekningsgrad_80_3_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .termindato(ETTER_WLB_1)
-            .antallBarn(3)
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(7)
-            .containsEntry(FELLESPERIODE, 90 + 280)
-            .containsEntry(FEDREKVOTE, 95)
-            .containsEntry(MØDREKVOTE, 95)
-            .containsEntry(FORELDREPENGER_FØR_FØDSEL, 15)
-            .containsEntry(FLERBARNSDAGER, 280)
-            .containsEntry(TILLEGG_FLERBARN, 280)
-            .containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-     Totale stønadskonto: - 59 uker (295 stønadsdager)
-     Stønadskonto fordeler seg slik:
-        - Fellesperiode: 20,2 uker (101 stønadsdager) + 57,4 uker (288 stønadsdager)
+        - Fellesperiode: 20,2 uker (101 stønadsdager) + 57,6 uker (288 stønadsdager)
         - Fedrekovte: 19 uker (95 stønadsdager)
         - Mødrekvote: 19 uker (95 stønadsdager)
         - Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
@@ -406,33 +322,6 @@ class StønadskontoRegelOrkestreringTest {
             .containsEntry(FLERBARNSDAGER, 288)
             .containsEntry(TILLEGG_FLERBARN, 288)
             .containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-        Totale stønadskonto: - 56 uker (280 stønadsdager)
-         Stønadskonto fordeler seg slik:
-        - Fellesperiode: 18 uker (90 stønadsdager) + 56 uker (280 stønadsdager)
-        - Fedrekovte: 19 uker (95 stønadsdager)
-        - Mødrekvote: 19 uker (95 stønadsdager)
-    */
-    @Test
-    void adopsjon_begge_har_rett_og_omsorg_dekningsgrad_80_3_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .omsorgsovertakelseDato(ETTER_WLB_1)
-            .antallBarn(3)
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(5)
-            .containsEntry(FELLESPERIODE, 90 + 280)
-            .containsEntry(FEDREKVOTE, 95)
-            .containsEntry(MØDREKVOTE, 95)
-            .containsEntry(FLERBARNSDAGER, 280)
-            .containsEntry(TILLEGG_FLERBARN, 280);
     }
 
     /*
@@ -520,36 +409,6 @@ class StønadskontoRegelOrkestreringTest {
     }
 
     /*
-    Totale stønadskonto: - 59 uker (295 stønadsdager)
-     Stønadskonto fordeler seg slik:
-    - Fellesperiode: 18 uker (90 stønadsdager) + 21 uker (105 stønadsdager)
-    - Fedrekovte: 19 uker (95 stønadsdager)
-    - Mødrekvote: 19 uker (95 stønadsdager)
-    - Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
-    */
-    @Test
-    void fødsel_begge_har_rett_og_omsorg_dekningsgrad_80_2_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(7)
-            .containsEntry(FELLESPERIODE, 90 + 105)
-            .containsEntry(FEDREKVOTE, 95)
-            .containsEntry(MØDREKVOTE, 95)
-            .containsEntry(FORELDREPENGER_FØR_FØDSEL, 15)
-            .containsEntry(FLERBARNSDAGER, 105)
-            .containsEntry(TILLEGG_FLERBARN, 105)
-            .containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
     Totale stønadskonto: - 61,2 uker (306 stønadsdager)
      Stønadskonto fordeler seg slik:
     - Fellesperiode: 20,2 uker (101 stønadsdager) + 21,2 uker (106 stønadsdager)
@@ -577,33 +436,6 @@ class StønadskontoRegelOrkestreringTest {
             .containsEntry(FLERBARNSDAGER, 106)
             .containsEntry(TILLEGG_FLERBARN, 106)
             .containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-        Totale stønadskonto: - 56 uker (280 stønadsdager)
-        Stønadskonto fordeler seg slik:
-        - Fellesperiode: 18 uker (90 stønadsdager) + 21 uker (105 stønadsdager)
-        - Fedrekovte: 19 uker (95 stønadsdager)
-        - Mødrekvote: 19 uker (95 stønadsdager)
-    */
-    @Test
-    void adopsjon_begge_har_rett_og_omsorg_dekningsgrad_80_2_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .omsorgsovertakelseDato(ETTER_WLB_1)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.BEGGE_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(5)
-            .containsEntry(FELLESPERIODE, 90 + 105)
-            .containsEntry(FEDREKVOTE, 95)
-            .containsEntry(MØDREKVOTE, 95)
-            .containsEntry(FLERBARNSDAGER, 105)
-            .containsEntry(TILLEGG_FLERBARN, 105);
     }
 
     /*
@@ -674,26 +506,6 @@ class StønadskontoRegelOrkestreringTest {
 
     /*
         Bare mor har rett til foreldrepenger.
-        Foreldrepenger: 56 uker (280 stønadsdager)
-        Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_mor_har_rett_begge_omsorg_dekningsgrad_80_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(1)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(2).containsEntry(FORELDREPENGER, 280).containsEntry(FORELDREPENGER_FØR_FØDSEL, 15);
-    }
-
-    /*
-        Bare mor har rett til foreldrepenger.
         Foreldrepenger: 58,2 uker (291 stønadsdager)
         Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
     */
@@ -710,25 +522,6 @@ class StønadskontoRegelOrkestreringTest {
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
         var stønadskontoer = stønadskontoResultat.getStønadskontoer();
         assertThat(stønadskontoer).hasSize(2).containsEntry(FORELDREPENGER, 291).containsEntry(FORELDREPENGER_FØR_FØDSEL, 15);
-    }
-
-    /*
-    Bare mor har rett til foreldrepenger.
-    Foreldrepenger: 56 uker (280 stønadsdager)
-    */
-    @Test
-    void adopsjon_bare_mor_har_rett_og_aleneomsorg_dekningsgrad_80_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .omsorgsovertakelseDato(ETTER_WLB_1)
-            .antallBarn(1)
-            .rettighetType(Rettighetstype.ALENEOMSORG)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(1).containsEntry(FORELDREPENGER, 280);
     }
 
     /*
@@ -793,27 +586,6 @@ class StønadskontoRegelOrkestreringTest {
 
     /*
         Mor har aleneomsorg og rett til foreldrepenger. 2 barn
-        Foreldrepenger: 56 uker (280 stønadsdager) + 21 uker (105 stønadsdager)
-        Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_mor_rett_dekningsgrad_80_2_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .termindato(ETTER_WLB_1)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 280 + 105).containsEntry(FORELDREPENGER_FØR_FØDSEL, 15)
-            .containsEntry(TILLEGG_FLERBARN, 105);
-    }
-
-    /*
-        Mor har aleneomsorg og rett til foreldrepenger. 2 barn
         Foreldrepenger: 58,2 uker (291 stønadsdager) + 21,2 uker (106 stønadsdager)
         Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
     */
@@ -831,26 +603,6 @@ class StønadskontoRegelOrkestreringTest {
         var stønadskontoer = stønadskontoResultat.getStønadskontoer();
         assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 291 + 106).containsEntry(FORELDREPENGER_FØR_FØDSEL, 15)
             .containsEntry(TILLEGG_FLERBARN, 106);
-    }
-
-    /*
-        Mor har aleneomsorg og rett til foreldrepenger. 2 barn
-        Foreldrepenger: 56 uker (280 stønadsdager) + 21 uker (105 stønadsdager)
-    */
-    @Test
-    void adopsjon_bare_mor_rett_og_aleneomsorg_dekningsgrad_80_2_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .omsorgsovertakelseDato(ETTER_WLB_1)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.ALENEOMSORG)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(2).containsEntry(FORELDREPENGER, 280 + 105)
-            .containsEntry(TILLEGG_FLERBARN, 105);
     }
 
     /*
@@ -916,27 +668,6 @@ class StønadskontoRegelOrkestreringTest {
 
     /*
     Mor har aleneomsorg og rett til foreldrepenger. 3 barn
-    Foreldrepenger: 56 uker (280 stønadsdager) + 56 uker (280 stønadsdager)
-    Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_mor_rett_dekningsgrad_80_3_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(3)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 280 + 280).containsEntry(FORELDREPENGER_FØR_FØDSEL, 15)
-            .containsEntry(TILLEGG_FLERBARN, 280);
-    }
-
-    /*
-    Mor har aleneomsorg og rett til foreldrepenger. 3 barn
     Foreldrepenger: 58,2 uker (291 stønadsdager) + 57,6 uker (288 stønadsdager)
     Foreldrepenger før fødsel: 3 uker (15 stønadsdager)
     */
@@ -954,25 +685,6 @@ class StønadskontoRegelOrkestreringTest {
         var stønadskontoer = stønadskontoResultat.getStønadskontoer();
         assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 291 + 288).containsEntry(FORELDREPENGER_FØR_FØDSEL, 15)
             .containsEntry(TILLEGG_FLERBARN, 288);
-    }
-
-    /*
-    Mor har aleneomsorg og rett til foreldrepenger. 3 barn
-    Foreldrepenger: 56 uker (280 stønadsdager) + 56 uker (280 stønadsdager)
-    */
-    @Test
-    void adopsjon_bare_mor_rett_og_aleneomsorg_dekningsgrad_80_3_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .omsorgsovertakelseDato(ETTER_WLB_1)
-            .antallBarn(3)
-            .rettighetType(Rettighetstype.ALENEOMSORG)
-            .brukerRolle(Brukerrolle.MOR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(2).containsEntry(FORELDREPENGER, 280 + 280).containsEntry(TILLEGG_FLERBARN, 280);
     }
 
     /*
@@ -1015,25 +727,6 @@ class StønadskontoRegelOrkestreringTest {
 
     /*
     Far har aleneomsorg og rett til  Foreldrepenger
-    Foreldrepenger: 56 uker (280 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_far_rett_og_aleneomsorg_dekningsgrad_80_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .termindato(ETTER_WLB_1)
-            .antallBarn(1)
-            .rettighetType(Rettighetstype.ALENEOMSORG)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(2).containsEntry(FORELDREPENGER, 280).containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-    Far har aleneomsorg og rett til  Foreldrepenger
     Foreldrepenger: 58,2 uker (291 stønadsdager)
     */
     @Test
@@ -1068,25 +761,6 @@ class StønadskontoRegelOrkestreringTest {
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
         var stønadskontoer = stønadskontoResultat.getStønadskontoer();
         assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 80 + 230 + 75 + 75).containsEntry(TILLEGG_FLERBARN, 230).containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-    Far har aleneomsorg og rett til Foreldrepenger. 3 barn
-    Foreldrepenger: 56 uker (280 stønadsdager) + 56 uker (280 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_far_rett_og_aleneomsorg_dekningsgrad_80_3_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(3)
-            .rettighetType(Rettighetstype.ALENEOMSORG)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 280 + 280).containsEntry(TILLEGG_FLERBARN, 280).containsEntry(FAR_RUNDT_FØDSEL, 10);
     }
 
     /*
@@ -1129,25 +803,6 @@ class StønadskontoRegelOrkestreringTest {
 
     /*
     Far har aleneomsorg og rett til Foreldrepenger. 2 barn
-    Foreldrepenger: 56 uker (280 stønadsdager) + 21 uker (105 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_far_rett_og_aleneomsorg_dekningsgrad_80_2_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.ALENEOMSORG)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 280 + 105).containsEntry(TILLEGG_FLERBARN, 105).containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-    Far har aleneomsorg og rett til Foreldrepenger. 2 barn
     Foreldrepenger: 58,2 uker (291 stønadsdager) + 21,2 uker (106 stønadsdager)
     */
     @Test
@@ -1167,25 +822,6 @@ class StønadskontoRegelOrkestreringTest {
 
     /*
        Far har rett til Foreldrepenger. 100% dekningsgrad.
-       Foreldrepenger: 40 uker (200 stønadsdager) med 8 uker minsterett
-    */
-    @Test
-    void fødsel_bare_far_rett_begge_omsorg_dekningsgrad_100_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .termindato(ETTER_WLB_1)
-            .antallBarn(1)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 50 + 75 + 75).containsEntry(BARE_FAR_RETT, 40).containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-       Far har rett til Foreldrepenger. 100% dekningsgrad.
        Foreldrepenger: 40 uker (200 stønadsdager) med 10 uker minsterett
     */
     @Test
@@ -1201,25 +837,6 @@ class StønadskontoRegelOrkestreringTest {
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
         var stønadskontoer = stønadskontoResultat.getStønadskontoer();
         assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 50 + 75 + 75).containsEntry(BARE_FAR_RETT, 50).containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-       Far har rett til Foreldrepenger. 80% dekningsgrad.
-       Foreldrepenger: 50 uker (250 stønadsdager) med 8 uker minsterett
-   */
-    @Test
-    void fødsel_bare_far_rett_begge_omsorg_dekningsgrad_80_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(1)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 250).containsEntry(BARE_FAR_RETT, 40).containsEntry(FAR_RUNDT_FØDSEL, 10);
     }
 
     /*
@@ -1245,26 +862,6 @@ class StønadskontoRegelOrkestreringTest {
        Far har rett til Foreldrepenger. 100% dekningsgrad. 3 barn.
        Foreldrepenger: 40 uker (200 stønadsdager) + 46 uker (230 stønadsdager)
        Flerbarnsdager: 46 uker (230 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_far_rett_begge_omsorg_dekningsgrad_100_3_barn() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .termindato(FØR_WLB)
-            .antallBarn(3)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 50 + 230 + 75 + 75).containsEntry(FLERBARNSDAGER, 230).containsEntry(TILLEGG_FLERBARN, 230);
-    }
-
-    /*
-       Far har rett til Foreldrepenger. 100% dekningsgrad. 3 barn.
-       Foreldrepenger: 40 uker (200 stønadsdager) + 46 uker (230 stønadsdager)
-       Flerbarnsdager: 46 uker (230 stønadsdager)
        Minsterett: 10 uker
     */
     @Test
@@ -1282,50 +879,6 @@ class StønadskontoRegelOrkestreringTest {
         assertThat(stønadskontoer).hasSize(4).containsEntry(FORELDREPENGER, 50 + 230 + 75 + 75)
             .containsEntry(TILLEGG_FLERBARN, 230)
             .containsEntry(BARE_FAR_RETT, 230)
-            .containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-    Far har rett til Foreldrepenger. 80% dekningsgrad. 3 barn.
-    Foreldrepenger: 50 uker (250 stønadsdager) + 56 uker (280 stønadsdager)
-    Flerbarnsdager: 56 uker (280 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_far_rett_begge_omsorg_dekningsgrad_80_3_barn() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(FØR_WLB)
-            .antallBarn(3)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 100 + 280 + 75 + 75).containsEntry(FLERBARNSDAGER, 280).containsEntry(TILLEGG_FLERBARN, 280);
-    }
-
-    /*
-    Far har rett til Foreldrepenger. 80% dekningsgrad. 3 barn.
-    Foreldrepenger: 50 uker (250 stønadsdager) + 56 uker (280 stønadsdager)
-    Minsterett: 56 uker (280 stønadsdager)
-    */
-    @Test
-    void wlb_fødsel_bare_far_rett_begge_omsorg_dekningsgrad_80_3_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(3)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(4)
-            .containsEntry(FORELDREPENGER, 100 + 280 + 75 + 75)
-            .containsEntry(TILLEGG_FLERBARN, 280)
-            .containsEntry(BARE_FAR_RETT, 280)
             .containsEntry(FAR_RUNDT_FØDSEL, 10);
     }
 
@@ -1355,28 +908,6 @@ class StønadskontoRegelOrkestreringTest {
     }
 
     /*
-       Far har rett til Foreldrepenger. 100% dekningsgrad. 2 barn.
-       Foreldrepenger: 40 uker (200 stønadsdager) + 17 uker (85 stønadsdager)
-       Flerbarnsdager: 17 uker (85 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_far_rett_begge_omsorg_dekningsgrad_100_2_barn() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(FØR_WLB)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 50 + 85 + 75 + 75)
-            .containsEntry(FLERBARNSDAGER, 85)
-            .containsEntry(TILLEGG_FLERBARN, 85);
-    }
-
-    /*
    Far har rett til Foreldrepenger. 100% dekningsgrad. 2 barn.
    Foreldrepenger: 40 uker (200 stønadsdager) + 17 uker (85 stønadsdager)
    Minsterett: 17 uker (85 stønadsdager)
@@ -1397,50 +928,6 @@ class StønadskontoRegelOrkestreringTest {
             .containsEntry(FORELDREPENGER, 50 + 85 + 75 + 75)
             .containsEntry(TILLEGG_FLERBARN, 85)
             .containsEntry(BARE_FAR_RETT, 85)
-            .containsEntry(FAR_RUNDT_FØDSEL, 10);
-    }
-
-    /*
-    Far har rett til Foreldrepenger. 80% dekningsgrad. 2 barn.
-    Foreldrepenger: 50 uker (250 stønadsdager) + 21 uker (105 stønadsdager)
-    Flerbarnsdager: 21 uker (105 stønadsdager)
-    */
-    @Test
-    void fødsel_bare_far_rett_begge_omsorg_dekningsgrad_80_2_barn() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(FØR_WLB)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(3).containsEntry(FORELDREPENGER, 100 + 105 + 75 + 75).containsEntry(FLERBARNSDAGER, 105).containsEntry(TILLEGG_FLERBARN, 105);
-    }
-
-    /*
-    Far har rett til Foreldrepenger. 80% dekningsgrad. 2 barn.
-    Foreldrepenger: 50 uker (250 stønadsdager) + 21 uker (105 stønadsdager)
-    Minsterett: 21 uker (105 stønadsdager)
-    */
-    @Test
-    void wlb_fødsel_bare_far_rett_begge_omsorg_dekningsgrad_80_2_barn_før_utvidelse() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(ETTER_WLB_1)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_80)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).hasSize(4)
-            .containsEntry(FORELDREPENGER, 100 + 105 + 75 + 75)
-            .containsEntry(TILLEGG_FLERBARN, 105)
-            .containsEntry(BARE_FAR_RETT, 105)
             .containsEntry(FAR_RUNDT_FØDSEL, 10);
     }
 
@@ -1512,23 +999,9 @@ class StønadskontoRegelOrkestreringTest {
             .build();
 
         assertThatCode(() -> stønadskontoRegelOrkestrering.beregnKontoer(grunnlag)).doesNotThrowAnyException();
+        var stønadskontoer = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag).getStønadskontoer();
+        assertThat(stønadskontoer).containsEntry(FELLESPERIODE, 80);
     }
-
-    @Test
-    void bfhr_flere_barn_ikke_minsterett_skal_gi_flerbarnsdager() {
-        var grunnlag = BeregnKontoerGrunnlag.builder()
-            .fødselsdato(FØR_WLB)
-            .antallBarn(2)
-            .rettighetType(Rettighetstype.BARE_SØKER_RETT)
-            .brukerRolle(Brukerrolle.FAR)
-            .dekningsgrad(Dekningsgrad.DEKNINGSGRAD_100)
-            .build();
-
-        var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var stønadskontoer = stønadskontoResultat.getStønadskontoer();
-        assertThat(stønadskontoer).containsKey(FLERBARNSDAGER);
-    }
-
 
     @Test
     void bfhr_flere_barn_minsterett_skal_ikke_gi_flerbarnsdager() {
